added new route group (generic) to separate general layout from pages like login, updated all imports wherever necessary, minor bug fixes
