+++ app/(generic)/input-forms/AddPatents.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import InputCard from "./InputCard";
-import { addPatent } from "../api/dbfunctions";
+import { addPatent } from "../../api/dbfunctions";
 import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
 
 const AddPatents = () => {
