+++ app/(generic)/my-journals/MyJournals.tsx
