+++ app/(generic)/my-journals/page.tsx
