+++ app/(generic)/report-gen/ReportPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ReportPage = () => {
+  return <div>Report</div>;
+};
+
+export default ReportPage;
